Move architecture specific files to designated directories
All architecture specific files can now be found in arch/arch_name
directories e.g. `arch/x86_64`.
