+++ LibOS/shim/src/arch/x86_64/.gitignore
@@ -1,0 +1,1 @@
+/asm-offsets.h
