+++ Pal/src/host/Linux/Makefile.am
@@ -1,6 +1,4 @@
 # Add host-specific compilation rules here
-SEC_DIR = security/$(PAL_HOST)
-
 CFLAGS += -fPIC -Wp,-U_FORTIFY_SOURCE -fno-builtin $(call cc-option,-Wtrampolines)
 
 CFLAGS += -Wextra $(call cc-option,-Wnull-dereference)
@@ -8,12 +6,12 @@ CFLAGS += -Wextra $(call cc-option,-Wnull-dereference)
 ASFLAGS += -DPIC -DSHARED -fPIC -DASSEMBLER -Wa,--noexecstack \
 	  -x assembler-with-cpp
 LDFLAGS += -shared -nostdlib -z combreloc -z defs \
-	  --version-script $(HOST_DIR)/pal.map -T $(HOST_DIR)/pal-$(ARCH).lds \
+	  --version-script $(HOST_DIR)/pal.map -T $(HOST_DIR)/arch/$(ARCH)/pal.lds \
 	  --eh-frame-hdr \
 	  -z relro -z now
 
 pal_loader = $(HOST_DIR)/libpal.so
 pal_lib = $(HOST_DIR)/libpal.so
-pal_lib_deps = pal-symbols $(HOST_DIR)/pal.map.template $(HOST_DIR)/pal-$(ARCH).lds
+pal_lib_deps = pal-symbols $(HOST_DIR)/pal.map.template $(HOST_DIR)/arch/$(ARCH)/pal.lds
 pal_lib_post =
 pal_static = $(HOST_DIR)/libpal.a
