+++ Pal/src/host/Skeleton/arch/x86_64/pal.lds
